Adding support for `grpc` library.
This required adding it as a dependency to the "clean"
environment (won't mean that on prod).

Also removed the installed `grpcio` from `lib/` (it
is unavoidable to add it, since that is how `google-cloud-language`
gets installed).

I checked (via `pipenv graph`) that no other dependencies were
unused after removing `grpcio` from `lib/`.